migrated all funcs to new models,views and js module.
Now all funcs of editor merged with js functionalities of Yura's sendToReview tool
