+++ client/css/main.styl
@@ -74,6 +74,7 @@ section.with_menu:hover > #rightcol
   align-content center
   border 2px solid #000000
   border-left 0
+
   a
     padding 30px 50px
     justify-content flex-start
