+++ client/js/editor/reviewModal.js
@@ -1,16 +1,19 @@
 import {editor} from './fabmain';
+import {disableControls} from './editor.js';
 
+require('./modals.js');
 const modal = document.getElementById('reviewModal');
 
 function sendToReview(event) {
     event.preventDefault();
+    let o = editor.canv.getActiveObject(),
+        g = editor.canv.getActiveGroup();
+        disableControls(o,g);
     const formData = new FormData(event.target);
     let imageReview = editor.canv.toJSON();
     // append image as base64 string
     formData.append('file', editor.canv.toDataURL("image/png", 1.0));
-    formData.append('file_json', editor.canv.toJSON());
-    console.log(formData);
-
+    formData.append('file_json', JSON.stringify(imageReview));
     fetch('/api/review',
         {
             method: 'POST',
@@ -18,13 +21,15 @@ function sendToReview(event) {
             body: formData
         }
     )
-        .then(function (response) {
-            if (response.status === 201) {
-                // show popup with success message here
-            }
-            else {
-                // show popup with error message here
-            }
+        .then((res) => res.json())
+        .then(function ({result}) {
+            document.getElementById('resulting').src = result.src;
+            document.getElementById('continue').href += result.rev;
+            document.getElementById('continue').style.display = "block";
+            document.getElementById('result_review').style.display = "block";
+        })
+        .catch(function (error) {
+            console.log('Request failed', error);
         });
     modal.style.display = "none";
 }
