+++ server/templates/admin/admin.html
@@ -12,7 +12,7 @@
                 <ul class="nav nav-pills nav-stacked" role="tablist">
                     <li><a href="{{ url_for ('backgrounds') }}">Фоны</a></li>
                     <li><a href="{{ url_for('users_page')}}">Пользователи</a></li>
-                    <li><a href="{{ url_for ('review_images') }}">Images in review</a></li>
+                    <li><a href="#">Images in review</a></li>
                 </ul>
             </div>
