+++ server/templates/editor_markuped.html
@@ -84,10 +84,6 @@
         <section class="container">
             {% block nav_edit %}
                 <nav class="navigation">
-                    <a href="#" id="to_send" data-bazooka="sendingReview"><i class="material-icons">done_all</i>
-                        Отправить на ревью</a>
-                    <a href="#"><i
-                            class="material-icons">insert_comment</i> Добавить коментарий для ревьюера</a>
                     <a href="#" id="result_review"><i class="material-icons">description</i>Просмотреть результат</a>
                     <a href="" id="continue"><i class="material-icons">flip_to_back</i>Продолжить редактировать</a>
                     <a href="#" class="special" id="download"><i class="material-icons">cloud_download</i> Скачать
@@ -122,8 +118,8 @@
         </div>
 
     </div>
-        {#  Include html for review modal window  #}
-        {% include 'editor/review_modal.html' %}
+    {#  Include html for review modal window  #}
+    {% include 'editor/review_modal.html' %}
 {% endblock %}
 {% block scripts %}
     <script src="{{ url_for('static', filename='editorbundle.js') }}"></script>
